changes to convert plugin contextualize to TS
