+++ packages/plugin-contextualize/rollup.config.mjs
@@ -1,0 +1,10 @@
+import createRollupConfig from "../../.rollup/index.mjs"
+
+const config = createRollupConfig({
+  input: './src/contextualize.ts',
+  external: [
+    '@bugsnag/core'
+  ]
+})
+
+export default config
