+++ packages/plugin-contextualize/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "rootDir": "src",
+    "outDir": "dist",
+    "declarationDir": "dist/types",
+    "lib": ["es2017"],
+    "types": ["node"]
+  },
+  "include": ["src/**/*.ts"]
+}
